LPD-102728 Move the widget page portlet tests to layout-admin-web
Playwright specs are grouped by the feature under test, and these three tests exercise adding, dragging, and removing widgets on a widget page, which is what widgetPage.spec.ts already covers. Living under portal-impl/portlet left them apart from the rest of the widget page coverage and attributed to the Frontend Other Testray component instead of Page Administration. The layout that the tests share is now created in a beforeEach scoped to the new describe block so it does not affect the tests already in the file.
